// src/types/student/Payment.types.ts
// types/student/payment.types.ts
export interface StudentPaymentPreview {
  AceYear?: number;
  AidedTypeId?: number;
  AidedTypes?: string;
  CCode?: number;
  ClassID?: number;
  ClassName?: string;
  Concession?: string;
  Email1?: string;
  EnrollMentDate?: string;
  EnrollMentNo?: string;
  FatherName?: string;
  FeeType?: string;
  FirstName?: string;
  Gender?: string;
  LastName?: string;
  MobileNo?: string;
  PaidFee?: number;
  PaymentDate?: string;
  SectionID?: number;
  TOTAL?: number;
  URNNO?: number;
}